requirements.txt
pandas
numpy
scikit-learn
requests
python-dotenv
matplotlib
